Moved the data to the raw data store location.
