+++ app/gamelib/src/data_store/asset_name.rs
@@ -20,13 +20,13 @@ impl AssetName {
     /// so each of the files in the ../data directory is included at compile time.
     pub const fn to_data(&self) -> &'static [u8] {
         match self {
-            AssetName::Cube => include_bytes!("../../../data/Cube.obj"),
-            AssetName::Suzanne => include_bytes!("../../../data/Suz.obj"),
-            AssetName::Triangle => include_bytes!("../../../data/Tri.obj"),
-            AssetName::CubeTexture => include_bytes!("../../../data/Cube.png"),
-            AssetName::TriangleTexture => include_bytes!("../../../data/Tri.png"),
-            AssetName::SuzanneTexture => include_bytes!("../../../data/Suz.png"),
-            AssetName::DemoMusic => include_bytes!("../../../data/DemoMusic.ogg"),
+            AssetName::Cube => include_bytes!("data/Cube.obj"),
+            AssetName::Suzanne => include_bytes!("data/Suz.obj"),
+            AssetName::Triangle => include_bytes!("data/Tri.obj"),
+            AssetName::CubeTexture => include_bytes!("data/Cube.png"),
+            AssetName::TriangleTexture => include_bytes!("data/Tri.png"),
+            AssetName::SuzanneTexture => include_bytes!("data/Suz.png"),
+            AssetName::DemoMusic => include_bytes!("data/DemoMusic.ogg"),
         }
     }
 }
