+++ app/src/rendering/textured_model.rs
@@ -1,6 +1,6 @@
 use super::indexed_model::IndexedModel;
 use super::texture::Texture;
-use gamelib::data_store::AssetName;
+use gamelib::data_store::asset_name::AssetName;
 use strum::IntoEnumIterator;
 use strum_macros::EnumIter;
