+++ app/src/target_tests/test_model_factory.rs
@@ -1,4 +1,4 @@
-use gamelib::data_store::AssetName;
+use gamelib::data_store::asset_name::AssetName;
 use crate::rendering::model_factory::ModelFactory;
 use crate::rendering::textured_model::TexturedModelName;
 use hashbrown::HashMap;
